新增:电枢转子测试仪V-2.2.180908-1330
新增:增加无线配置界面
新增:增加电阻定档采集
新增:增加匝间多次采集
新增:增加单工位与双工位
新增:增加启动时清理数据
新增:增加按型号导出数据
新增:增加下位机异常提示
新增:增加退出时保存提示
新增:增加读取1302时间功能
新增:增加显示测试时间功能
新增:增加电阻采样取消功能
新增:增加匝间波形自动放大功能
新增:增加匝间快速设置间隔功能
新增:增加测试界面结果放大显示
新增:增加权限管理超级管理员选择
新增:系统设置参数变化时加入提示音

修复:部分界面修改
修复:修正耐压异常报警
修复:修正启动中拉滑罩死机
修复:修正新建型号越来越大的错误

+++ app/boxdouble.cpp
@@ -28,7 +28,7 @@ QWidget *BoxDouble::createEditor(QWidget *parent,
 
 void BoxDouble::setEditorData(QWidget *editor, const QModelIndex &index) const
 {
-    int value = index.model()->data(index, Qt::EditRole).toInt();
+    double value = index.model()->data(index, Qt::EditRole).toDouble();
     QDoubleSpinBox *spinBox = static_cast<QDoubleSpinBox*>(editor);
     spinBox->setValue(value);
 }
